Bugfix: file size is `uint64_t`, not `uint32_t`.
All tests pass with GCC (but not with Clang).
All tests pass with GCC (but not with Clang).
All tests pass with GCC (but not with Clang).
All tests pass with GCC (but not with Clang).

	modified:   code/ylikuutio/loaders/font_loader.cpp
	modified:   code/ylikuutio/tests/googletest/test_font_loader.cpp

+++ code/ylikuutio/loaders/font_loader.cpp
@@ -315,7 +315,7 @@ namespace loaders
             std::vector<std::string>& unicode_strings)
     {
         std::string file_content = file::slurp(font_file_path);
-        const uint32_t file_size = file_content.size();
+        const uint64_t file_size = file_content.size();
         char* SVG_data = new char[file_size];
 
         if (SVG_data == nullptr)
